refactor: move all error imports a single file
