+++ src/presentation/controllers/singup.spec.ts
@@ -1,7 +1,6 @@
 import { SingUpController } from './singup'
-import { MissingParamError } from '../errors/missing-param-error'
+import { MissingParamError, InvalidParamError } from '../errors'
 import { EmailValidator } from '../protocols/email-validator'
-import { InvalidParamError } from '../errors/invalid-param-error'
 import { ServerError } from '../errors/server-error'
 
 interface SutTypes {
