+++ src/presentation/controllers/singup.ts
@@ -1,9 +1,8 @@
 import { Controller } from './../protocols/controller'
 import { HttpRequest, HttpResponse } from '../protocols/http'
-import { MissingParamError } from '../errors/missing-param-error'
+import { MissingParamError, InvalidParamError } from '../errors'
 import { badRequest, serverError } from '../helpers/http-helper'
 import { EmailValidator } from '../protocols/email-validator'
-import { InvalidParamError } from '../errors/invalid-param-error'
 
 export class SingUpController implements Controller {
   private readonly emailValidator: EmailValidator
